src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-evaluation>=1.5.0",
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects>=1.0.0b9",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-monitor-events-extension>=0.1.0",
    "azure-monitor-opentelemetry>=1.6.8",
    "azure-search-documents>=11.5.2",
    "fastapi>=0.115.12",
    "openai>=1.75.0",
    "opentelemetry-api>=1.31.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
    "opentelemetry-exporter-otlp-proto-http>=1.31.1",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "opentelemetry-instrumentation-openai>=0.39.2",
    "opentelemetry-sdk>=1.31.1",
    "pytest>=8.2,<9",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "semantic-kernel>=1.28.1",
    "uvicorn>=0.34.2",
]
